Split components into separate files. Removed headless UI wrapper.
