+++ src/ui/components/ProtocolViewer/components/ProtocolRequestDetail.module.css
@@ -1,0 +1,74 @@
+.AccordionHeader {
+  font-size: 1rem;
+  line-height: 1rem;
+  flex: 0 0 auto;
+  padding: 0.5rem 0;
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+  border-bottom: 1px solid var(--theme-border);
+  gap: 0.25rem;
+  user-select: none;
+}
+
+.ColorErrored {
+  color: var(--console-error-color);
+}
+.ColorPending {
+  color: var(--light-grey);
+}
+
+.DetailPanelHeader {
+  font-size: 0.9rem;
+  line-height: 0.9rem;
+  padding: 0.5rem 0;
+}
+
+.DetailPanelHeaderPrimary {
+  flex: 1 1 auto;
+  text-overflow: ellipsis;
+  white-space: pre;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 1ch;
+}
+
+.DetailPanelHeaderSecondary {
+  flex: 0 0 auto;
+  padding-right: 0.5rem;
+}
+
+.BugReportLink,
+.HoneycombLink {
+  color: white;
+  border-radius: 0.25rem;
+  width: 1.25rem;
+  height: 1.25rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.BugReportLink {
+  background-color: var(--red-50);
+}
+
+.HoneycombLink {
+  background-color: var(--blue-50);
+}
+
+.DetailPanel {
+  padding: 0 0.5rem;
+  border-bottom: 1px solid var(--theme-border);
+}
+:root:global(.theme-dark) .DetailPanel {
+  background-color: #232938;
+}
+:root:global(.theme-light) .DetailPanel {
+  background-color: var(--grey-20);
+}
+
+.JSONViewerContainer {
+  margin: 0.5rem 0;
+}
